+++ quartz/src/jvmAndroid/kotlin/com/vitorpamplona/quartz/nip01Core/cache/EventInterner.jvmAndroid.kt
@@ -18,8 +18,10 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.quartz.nip01Core.core
+package com.vitorpamplona.quartz.nip01Core.cache
 
+import com.vitorpamplona.quartz.nip01Core.core.Event
+import com.vitorpamplona.quartz.nip01Core.core.HexKey
 import java.lang.ref.WeakReference
 import java.util.concurrent.ConcurrentHashMap
